+++ app/mailers/application_mailer.rb
@@ -4,11 +4,12 @@
 class ApplicationMailer < ActionMailer::Base
   default from: "#{ENV['website_name']} <#{ActionMailer::Base.smtp_settings[:email]}>"
   add_template_helper(ApplicationHelper)
+  add_template_helper(EmailHelper)
   layout 'mailer'
 
   protected
 
   def setup_email
-    # attachments.inline['nsrr-logo.png'] = File.read('app/assets/images/nsrr_logo_64.png') rescue nil
+    # attachments.inline['tasktracker-logo.png'] = File.read('app/assets/images/tasktracker_logo.png') rescue nil
   end
 end